fix compiler errors

+++ amatracker/Movement Detail View Controller/Extensions/MovementDetailVC+TableViewDelegate.swift
@@ -4,19 +4,23 @@ import UIKit
 extension MovementDetailViewController : UITableViewDelegate {
      
     func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
-        let record = records[indexPath.row]
-        if(record.numberOfReps == 1) {
-            performSegue(withIdentifier: SHOW_PERCENTAGES_SEGUE, sender: nil)
+        if let record = viewModel.recordViewModel(for: indexPath.row)?.record {
+            if(record.numberOfReps == 1) {
+                performSegue(withIdentifier: SHOW_PERCENTAGES_SEGUE, sender: nil)
+            }
         }
+       
     }
     
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == UITableViewCellEditingStyle.delete {
-            let recordToDelete: Record = records[indexPath.row]
-            recordService.deleteRecord(recordToDelete)
+            if let recordToDelete: Record = viewModel.recordViewModel(for: indexPath.row)?.record {
+                recordService.deleteRecord(recordToDelete)
+                viewModel.removeRecordAt(index: indexPath.row)
+                
+                tableView.deleteRows(at: [indexPath], with: UITableViewRowAnimation.automatic)
+            }
             
-            records.remove(at: indexPath.row)
-            tableView.deleteRows(at: [indexPath], with: UITableViewRowAnimation.automatic)
         }
     }
 }
